1. Changed the env variable name from 'ALEMBIC_DB_URL' to 'DATABASE_URL', 2. Deleted out the commented unused code for CRUD, 3. Renamed the path operations for DELETE, changed them to PUT, renamed to activate_* and deactivate_*, 4. The controller CRUD functions are split into 4 different files within the new controllers module, After this i will get on creating a more granular project structure
