+++ app/controllers/crud.py
@@ -1,0 +1,7 @@
+from sqlalchemy.orm import Session
+from app import models, schemas
+
+from .users_crud import *
+from .user_types_crud import *
+from .courses_crud import *
+from .enrollments_crud import *
